requirements.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
flask==3.1.2
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
werkzeug==3.1.3
